Cleanup sequence plugin registrations
