+++ plugins/sequence/src/GCContentAdapter/index.ts
@@ -1,9 +1,22 @@
 import PluginManager from '@jbrowse/core/PluginManager'
+import AdapterType from '@jbrowse/core/pluggableElementTypes/AdapterType'
+
 import configSchemaF from './configSchema'
 
 export default (pluginManager: PluginManager) => {
-  return {
-    configSchema: pluginManager.load(configSchemaF),
-    getAdapterClass: () => import('./GCContentAdapter').then(r => r.default),
-  }
+  pluginManager.addAdapterType(
+    () =>
+      new AdapterType({
+        name: 'GCContentAdapter',
+        adapterMetadata: {
+          category: null,
+          hiddenFromGUI: true,
+          displayName: null,
+          description: null,
+        },
+        configSchema: configSchemaF(pluginManager),
+        getAdapterClass: () =>
+          import('./GCContentAdapter').then(r => r.default),
+      }),
+  )
 }
