+++ plugins/sequence/src/IndexedFastaAdapter/index.ts
@@ -1,1 +1,22 @@
-export { default as configSchema } from './configSchema'
+import PluginManager from '@jbrowse/core/PluginManager'
+import AdapterType from '@jbrowse/core/pluggableElementTypes/AdapterType'
+
+import configSchema from './configSchema'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addAdapterType(
+    () =>
+      new AdapterType({
+        name: 'IndexedFastaAdapter',
+        configSchema,
+        adapterMetadata: {
+          category: null,
+          hiddenFromGUI: true,
+          displayName: null,
+          description: null,
+        },
+        getAdapterClass: () =>
+          import('./IndexedFastaAdapter').then(r => r.default),
+      }),
+  )
+}
