+++ plugins/sequence/src/ReferenceSequenceTrack/configSchema.ts
